# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "interacting with smart contract, collecting tokens, perform trade",
        "title": "Flash Loan Arbitrage bot",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/parser/pools": {
            "get": {
                "description": "Save pools from parser to storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parse"
                ],
                "summary": "Store parsed pools",
                "operationId": "ParseWrite",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ListPools"
                        }
                    },
                    "507": {
                        "description": "Insufficient Storage",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/provider/tokens": {
            "get": {
                "description": "Request token list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tradecase"
                ],
                "summary": "List Tokens",
                "operationId": "listTokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TokenList"
                        }
                    }
                }
            },
            "post": {
                "description": "Add list of tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tradecase"
                ],
                "summary": "Add Tokens",
                "operationId": "addTokens",
                "parameters": [
                    {
                        "description": "Add tokens",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TokenList"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/TokenList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/storage/pools": {
            "get": {
                "description": "Get list of all pools from storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parsecase",
                    "tradecase"
                ],
                "summary": "Get Pools",
                "operationId": "getPoolList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ListPools"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "507": {
                        "description": "Insufficient Storage",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add list of pools to storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parsecase",
                    "tradecase"
                ],
                "summary": "Add Pools",
                "operationId": "addPools",
                "parameters": [
                    {
                        "description": "Add pools",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ListPools"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ListPools"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "507": {
                        "description": "Insufficient Storage",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete pools from storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parsecase",
                    "tradecase"
                ],
                "summary": "Delete Pools",
                "operationId": "deletePools",
                "parameters": [
                    {
                        "description": "Delete pools",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ListPools"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ListPools"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "507": {
                        "description": "Insufficient Storage",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/storage/tokens": {
            "get": {
                "description": "Get list of all tokens from storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parsecase",
                    "tradecase"
                ],
                "summary": "Get Tokens",
                "operationId": "getTokensList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TokenList"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "507": {
                        "description": "Insufficient Storage",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add list of tokens to storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parsecase",
                    "tradecase"
                ],
                "summary": "Add Tokens",
                "operationId": "storeTokens",
                "parameters": [
                    {
                        "description": "Add tokens",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TokenList"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/TokenList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "507": {
                        "description": "Insufficient Storage",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete tokens from storage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parsecase",
                    "tradecase"
                ],
                "summary": "Delete Tokens",
                "operationId": "deleteTokens",
                "parameters": [
                    {
                        "description": "Delete tokens",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TokenList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TokenList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "507": {
                        "description": "Insufficient Storage",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ErrorResponse": {
            "description": "Error response object",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
...
        "ListPools": {
            "description": "Request list of trading pools",
            "type": "object",
            "properties": {
                "pools": {
                    "description": "list of trade pools",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Pool"
                    }
                }
            }
        },
...
        "TokenList": {
            "description": "Request list of tokens",
            "type": "object",
            "properties": {
                "tokens": {
                    "description": "list of tokens",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Token"
                    }
                }
            }
        },
        "entities.Pool": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "pair": {
                    "$ref": "#/definitions/entities.TokenPair"
                },
                "protocol": {
                    "$ref": "#/definitions/entities.SwapProtocol"
                }
            }
        },
        "entities.SwapProtocol": {
            "type": "object",
            "properties": {
                "factory": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "router": {
                    "type": "string"
                }
            }
        },
        "entities.Token": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "wei": {
                    "type": "integer"
                }
            }
        },
        "entities.TokenPair": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "token0": {
                    "$ref": "#/definitions/entities.Token"
                },
                "token1": {
                    "$ref": "#/definitions/entities.Token"
                }
            }
        },
...
    }
}